utils: fix MSVC build failures for windows-x86-64
The windows-build probe failed with 497 C4273/C2491 dllimport errors
plus C4053 void-expression diagnostics, blocking the windows-x86-64
release leg (v0.1.11 tag run 33982029953).

Root causes fixed:

- jsonrpc_helpers.h declared AIRY_API functions inside a WIN32 block
  without AIRY_BUILDING_DLL, hijacking the include chain into
  dllimport mode for every consumer (C4273 x497). Include <compat.h>
  instead, matching platform_base.h and logging.h.

- export.h defined dllimport when AIRY_BUILDING_DLL was unset,
  breaking consumers that link commons statically on Windows; leave
  AIRY_API empty to match <compat.h>.

- platform_misc.h redeclared airy_strlcpy/strlcat without AIRY_API,
  colliding with <compat.h> declarations (C4273).

- atomic_compat_api.h store macros ended in void-returning
  atomic_store_N branches; MSVC rejects void operands in ?: chains
  (C4053, promoted to errors by /WX). Wrap each branch to yield int.

- gethostname shim used size_t for the length parameter; the CRT
  prototype takes int (C4028).

- Hand-written (volatile _Atomic ...) casts expanded to
  "volatile volatile" under MSVC's _Atomic definition (C4114).

- __builtin_* calls are the GCC-only escape hatch for STRICT-mode
  poisoned functions; MSVC has neither poison nor __builtin_*.
  Introduce the AIRY_IMPL_* backend macro layer in
  airy_memory_inline.h (GCC/Clang keep __builtin_*, MSVC maps to
  CRT) and convert AIRY_STRNCPY_TERM/AIRY_MEMCPY_SAFE/AIRY_MEMSET/
  AIRY_MEMCPY/AIRY_MEMMOVE and direct call sites to use it.

- Diagnostic pragma push/pop is GCC/Clang-only; guard it (C4068).

Verified: full GCC build in STRICT compliance mode (poison enabled)
passes 100% with no new diagnostics.

+++ utils/compat/compat.c
@@ -199,7 +199,7 @@ const char *airy_build_info(void)
 #include <windows.h>
 #include "airy_memory.h"
 
-int gethostname(char *name, size_t len)
+int gethostname(char *name, int len)
 {
     DWORD size = (DWORD)len;
     if (!GetComputerNameA(name, &size)) {
